build: Updating mediawiki/mediawiki-phan-config to 0.9.0
Depends-On: I9661ed8dd80cb827d7a1414c1eef952c0933a1f0
Depends-On: I719fec91cb765a1673299974ec0eea59c37ce72f

+++ includes/EventLogging.php
@@ -134,7 +134,7 @@ class EventLogging {
 	 * @return bool True if the event should be included (sampled in), false if not (sampled out)
 	 */
 	public static function sessionInSample( $populationSize, $sessionId ) {
-		$decimal = base_convert( substr( $sessionId, 0, 8 ), 16, 10 );
+		$decimal = (int)base_convert( substr( $sessionId, 0, 8 ), 16, 10 );
 		return $decimal % $populationSize === 0;
 	}
 }
